14: Order suggestions according to their similarity to the search query - Move files
