+++ auction-backend/src/main/java/com/atlantbh/internship/auction/app/service/featured/criteria/RegularUserCriteria.java
@@ -2,7 +2,7 @@ package com.atlantbh.internship.auction.app.service.featured.criteria;
 
 import com.atlantbh.internship.auction.app.builder.SpecificationBuilder;
 import com.atlantbh.internship.auction.app.entity.Item;
-import com.atlantbh.internship.auction.app.service.specification.ItemSpecification;
+import com.atlantbh.internship.auction.app.specification.ItemSpecification;
 import org.springframework.data.jpa.domain.Specification;
 
 import java.time.ZonedDateTime;
