+++ auction-backend/src/main/java/com/atlantbh/internship/auction/app/specification/ItemSpecification.java
@@ -1,4 +1,4 @@
-package com.atlantbh.internship.auction.app.service.specification;
+package com.atlantbh.internship.auction.app.specification;
 
 import com.atlantbh.internship.auction.app.entity.Bid;
 import com.atlantbh.internship.auction.app.entity.Category;
