+++ auction-backend/src/main/java/com/atlantbh/internship/auction/app/specification/ItemSpecificationProcessor.java
@@ -1,4 +1,4 @@
-package com.atlantbh.internship.auction.app.service.specification;
+package com.atlantbh.internship.auction.app.specification;
 
 import com.atlantbh.internship.auction.app.dto.item.requests.GetItemsRequest;
 import com.atlantbh.internship.auction.app.entity.Item;
